refactor(cli): share env-assembly and signal-teardown between shell and run
`coi shell` and `coi run` had two byte-identical blocks duplicated (a real
drift risk — the review's cli #1 finding):

- The config-env middle (defaults.environment -> forward_env with the identical
  unset-var warning -> env_commands, last-wins) -> new `(*App).applyConfigEnv`.
  buildContainerEnv/appendEnvArgs keep their path-specific baseline + emit.
- The SIGINT/SIGTERM handler goroutine + `pipeline.Run(ctx, ...)` ->
  new `runPipelineWithSignals`. Each caller keeps its own `pipeline :=` and
  `defer pipeline.Teardown()` (idempotent), so behavior is unchanged.

Both helpers live in cli/common.go. Same precedence, same warning text, same
fatal-on-env-command-error, same signal semantics. No behavior change.

Left the slot-allocation duplication alone for now: shell and run print
intentionally different reuse messages ("Reusing stopped persistent container"
vs "Reusing persistent container"), so a shared helper would change output.

go build/vet, go test ./... (23 pkgs), golangci-lint 0 issues, gofmt clean.
