+++ internal/cli/common.go
@@ -1,10 +1,14 @@
 package cli
 
 import (
+	"context"
 	"fmt"
 	"os"
+	"os/signal"
+	"syscall"
 
 	"github.com/mensfeld/code-on-incus/internal/alias"
+	"github.com/mensfeld/code-on-incus/internal/session"
 )
 
 // confirmYN prints prompt and reads a line from stdin, returning true only when
@@ -42,4 +46,57 @@ func resolveNameOrAlias(nameOrAlias string) (string, error) {
 // func(string): it writes msg to stderr with a trailing newline.
 func stderrLogFn(msg string) {
 	fmt.Fprintf(os.Stderr, "%s\n", msg)
+}
+
+// applyConfigEnv overlays the three config-sourced environment layers onto env
+// in increasing-priority (last-wins) order: [defaults].environment, then
+// forward_env (host values, warning on an unset var), then env_commands
+// (freshly minted per session, fatal on error). Shared by coi shell
+// (buildContainerEnv) and coi run (appendEnvArgs) so the precedence and the
+// unset-var warning can't drift between the two launch paths.
+func (a *App) applyConfigEnv(env map[string]string) error {
+	// Static environment from config (defaults.environment + profile environment)
+	for k, v := range a.cfg.Defaults.Environment {
+		env[k] = v
+	}
+	// Resolve forward_env from config, look up host values
+	for _, name := range a.cfg.Defaults.ForwardEnv {
+		if val, ok := os.LookupEnv(name); ok {
+			env[name] = val
+		} else {
+			fmt.Fprintf(os.Stderr, "Warning: forward_env variable %q is not set on host, skipping\n", name)
+		}
+	}
+	// Command-sourced env vars (highest precedence — freshly minted per session).
+	envCommandValues, err := a.resolveEnvCommands()
+	if err != nil {
+		return err
+	}
+	for k, v := range envCommandValues {
+		env[k] = v
+	}
+	return nil
+}
+
+// runPipelineWithSignals installs a SIGINT/SIGTERM handler that triggers
+// pipeline.Teardown immediately — so cleanup runs even while a blocking incus
+// exec ignores ctx cancellation — then runs the pipeline. A dedicated sigChan
+// (not ctx.Done) avoids the non-determinism of both firing at once; Teardown is
+// idempotent so the caller's own `defer pipeline.Teardown()` stays safe. Shared
+// by coi shell and coi run.
+func runPipelineWithSignals(ctx context.Context, pipeline *session.Pipeline, phases ...session.Phase) error {
+	sigChan := make(chan os.Signal, 1)
+	signal.Notify(sigChan, os.Interrupt, syscall.SIGTERM)
+	defer signal.Stop(sigChan)
+	done := make(chan struct{})
+	defer close(done)
+	go func() {
+		select {
+		case <-sigChan:
+			fmt.Fprintf(os.Stderr, "\nReceived interrupt signal, cleaning up...\n")
+			pipeline.Teardown()
+		case <-done:
+		}
+	}()
+	return pipeline.Run(ctx, phases...)
 }
